Reorganizing unit test
