+++ .travis.yml
@@ -1,13 +1,12 @@
 language: php
 php:
+  - "7.1"
+  - "7.0"
   - "5.6"
-  - "5.5"
-  - "5.4"
 
 install:
   - composer install
 
 script: 
-  - cd tests
-  - phpunit --bootstrap bootstrap.php . 
+  - phpunit
