--- nbproject/project.properties
@@ -1,21 +1,0 @@
-auxiliary.org-netbeans-modules-php-phpunit.bootstrap_2e_create_2e_tests=true
-auxiliary.org-netbeans-modules-php-phpunit.bootstrap_2e_enabled=true
-auxiliary.org-netbeans-modules-php-phpunit.bootstrap_2e_path=tests/bootstrap.php
-auxiliary.org-netbeans-modules-php-phpunit.configuration_2e_enabled=false
-auxiliary.org-netbeans-modules-php-phpunit.configuration_2e_path=
-auxiliary.org-netbeans-modules-php-phpunit.customSuite_2e_enabled=false
-auxiliary.org-netbeans-modules-php-phpunit.customSuite_2e_path=
-auxiliary.org-netbeans-modules-php-phpunit.phpUnit_2e_enabled=false
-auxiliary.org-netbeans-modules-php-phpunit.phpUnit_2e_path=
-auxiliary.org-netbeans-modules-php-phpunit.test_2e_groups_2e_ask=false
-auxiliary.org-netbeans-modules-php-phpunit.test_2e_run_2e_all=false
-file.reference.xmlutil-tests=tests
-include.path=${php.global.include.path}
-php.version=PHP_54
-source.encoding=UTF-8
-src.dir=.
-tags.asp=false
-tags.short=false
-test.src.dir=${file.reference.xmlutil-tests}
-testing.providers=PhpUnit
-web.root=.
