+++ tests/XmlUtilTest.php
@@ -1,17 +1,21 @@
 <?php
 
-namespace ByJG\Util;
+namespace Tests;
 
-/**
- * Generated by PHPUnit_SkeletonGenerator on 2015-07-26 at 15:17:22.
- */
-class XmlUtilTest extends \PHPUnit_Framework_TestCase
+use ByJG\Util\XmlUtil;
+
+// backward compatibility
+if (!class_exists('\PHPUnit\Framework\TestCase')) {
+    class_alias('\PHPUnit_Framework_TestCase', '\PHPUnit\Framework\TestCase');
+}
+
+class XmlUtilTest extends \PHPUnit\Framework\TestCase
 {
 
     const XMLHEADER = '<?xml version="1.0" encoding="utf-8"?>';
 
     /**
-     * @var XmlUtil
+     * @var \ByJG\Util\XmlUtil
      */
     protected $object;
 
@@ -30,30 +34,21 @@ class XmlUtilTest extends \PHPUnit_Framework_TestCase
      */
     protected function tearDown()
     {
-        
+
     }
 
-    /**
-     * @covers ByJG\Util\XmlUtil::createXmlDocument
-     */
     public function testCreateXmlDocument()
     {
         $xml = XmlUtil::createXmlDocument();
         $this->assertEquals(self::XMLHEADER . "\n", $xml->saveXml());
     }
 
-    /**
-     * @covers ByJG\Util\XmlUtil::createXmlDocumentFromFile
-     */
     public function testCreateXmlDocumentFromFile()
     {
         $xml = XmlUtil::createXmlDocumentFromFile(__DIR__ . '/buggy.xml');
         $this->assertEquals(self::XMLHEADER . "\n<root><node><subnode>value</subnode></node></root>\n", $xml->saveXML());
     }
 
-    /**
-     * @covers ByJG\Util\XmlUtil::createXmlDocumentFromStr
-     */
     public function testCreateXmlDocumentFromStr()
     {
         $xmlStr = '<root/>';
@@ -61,9 +56,6 @@ class XmlUtilTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(self::XMLHEADER . "\n<root/>\n", $xml->saveXML());
     }
 
-    /**
-     * @covers ByJG\Util\XmlUtil::createDocumentFromNode
-     */
     public function testCreateDocumentFromNode()
     {
         $xml = XmlUtil::createXmlDocumentFromFile(__DIR__ . '/buggy.xml');
@@ -74,9 +66,6 @@ class XmlUtilTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(self::XMLHEADER . "\n<subnode>value</subnode>\n", $xmlFinal->saveXML());
     }
 
-    /**
-     * @covers ByJG\Util\XmlUtil::fixXmlHeader
-     */
     public function testFixXmlHeader1()
     {
         $xml = '<root/>';
@@ -84,9 +73,6 @@ class XmlUtilTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(self::XMLHEADER . '<root/>', $result);
     }
 
-    /**
-     * @covers ByJG\Util\XmlUtil::fixXmlHeader
-     */
     public function testFixXmlHeader2()
     {
         $xml = '<?xml?><root/>';
@@ -94,9 +80,6 @@ class XmlUtilTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(self::XMLHEADER . '<root/>', $result);
     }
 
-    /**
-     * @covers ByJG\Util\XmlUtil::fixXmlHeader
-     */
     public function testFixXmlHeader3()
     {
         $xml = '<?xml version="1.0"?><root/>';
@@ -104,9 +87,6 @@ class XmlUtilTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(self::XMLHEADER . '<root/>', $result);
     }
 
-    /**
-     * @covers ByJG\Util\XmlUtil::fixXmlHeader
-     */
     public function testFixXmlHeader4()
     {
         $xml = '<?xml encoding="utf8"?><root/>';
@@ -114,9 +94,6 @@ class XmlUtilTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(self::XMLHEADER . '<root/>', $result);
     }
 
-    /**
-     * @covers ByJG\Util\XmlUtil::fixXmlHeader
-     */
     public function testFixXmlHeader5()
     {
         $xml = '<?xml encoding="ascii"?><root/>';
@@ -124,9 +101,6 @@ class XmlUtilTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(self::XMLHEADER . '<root/>', $result);
     }
 
-    /**
-     * @covers ByJG\Util\XmlUtil::saveXmlDocument
-     */
     public function testSaveXmlDocument()
     {
         $filename = sys_get_temp_dir() . '/save.xml';
@@ -146,10 +120,6 @@ class XmlUtilTest extends \PHPUnit_Framework_TestCase
         unlink($filename);
     }
 
-    /**
-     * @covers ByJG\Util\XmlUtil::getFormattedDocument
-     * @todo   Implement testGetFormattedDocument().
-     */
     public function testGetFormattedDocument()
     {
         $xml = XmlUtil::createXmlDocumentFromFile(__DIR__ . '/buggy.xml');
@@ -159,14 +129,11 @@ class XmlUtilTest extends \PHPUnit_Framework_TestCase
 
         $this->assertTrue($xml->preserveWhiteSpace);
         $this->assertFalse($xml->formatOutput);
-        $this->assertEquals(self::XMLHEADER . "\n<root>\n  <node>\n    <subnode>value</subnode>\n  </node>\n</root>\n",
-            $formatted);
+        $this->assertEquals(
+            self::XMLHEADER . "\n<root>\n  <node>\n    <subnode>value</subnode>\n  </node>\n</root>\n" , $formatted
+        );
     }
 
-    /**
-     * @covers ByJG\Util\XmlUtil::addNamespaceToDocument
-     * @todo   Implement testAddNamespaceToDocument().
-     */
     public function testAddNamespaceToDocument()
     {
         // Remove the following lines when you implement this test.
@@ -175,10 +142,6 @@ class XmlUtilTest extends \PHPUnit_Framework_TestCase
         );
     }
 
-    /**
-     * @covers ByJG\Util\XmlUtil::addNodeFromFile
-     * @todo   Implement testAddNodeFromFile().
-     */
     public function testAddNodeFromFile()
     {
         // Remove the following lines when you implement this test.
@@ -187,10 +150,6 @@ class XmlUtilTest extends \PHPUnit_Framework_TestCase
         );
     }
 
-    /**
-     * @covers ByJG\Util\XmlUtil::addNodeFromNode
-     * @todo   Implement testAddNodeFromNode().
-     */
     public function testAddNodeFromNode()
     {
         // Remove the following lines when you implement this test.
@@ -199,11 +158,6 @@ class XmlUtilTest extends \PHPUnit_Framework_TestCase
         );
     }
 
-    /**
-     * @covers ByJG\Util\XmlUtil::createChild
-     * @covers ByJG\Util\XmlUtil::createChildBefore
-     * @covers ByJG\Util\XmlUtil::createChildBeforeNode
-     */
     public function testCreateChild()
     {
         $dom = XmlUtil::createXmlDocumentFromStr('<root/>');
@@ -255,9 +209,6 @@ class XmlUtilTest extends \PHPUnit_Framework_TestCase
         );
     }
 
-    /**
-     * @covers ByJG\Util\XmlUtil::addTextNode
-     */
     public function testAddTextNode()
     {
         $dom = XmlUtil::createXmlDocumentFromStr('<root><subject></subject></root>');
@@ -277,9 +228,6 @@ class XmlUtilTest extends \PHPUnit_Framework_TestCase
         );
     }
 
-    /**
-     * @covers ByJG\Util\XmlUtil::addAttribute
-     */
     public function testAddAttribute()
     {
         $dom = XmlUtil::createXmlDocumentFromStr('<root><subject>Text</subject></root>');
@@ -299,10 +247,6 @@ class XmlUtilTest extends \PHPUnit_Framework_TestCase
         );
     }
 
-    /**
-     * @covers ByJG\Util\XmlUtil::selectNodes
-     * @covers ByJG\Util\XmlUtil::selectSingleNode
-     */
     public function testSelectNodes()
     {
         $dom = XmlUtil::createXmlDocumentFromStr('<root><a><item arg="1"/><item arg="2"><b1/><b2/></item><item arg="3"/></a></root>');
@@ -320,10 +264,6 @@ class XmlUtilTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('b2', $node->nodeName);
     }
 
-    /**
-     * @covers ByJG\Util\XmlUtil::registerNamespaceForFilter
-     * @todo   Implement testRegisterNamespaceForFilter().
-     */
     public function testRegisterNamespaceForFilter()
     {
         // Remove the following lines when you implement this test.
@@ -332,10 +272,6 @@ class XmlUtilTest extends \PHPUnit_Framework_TestCase
         );
     }
 
-    /**
-     * @covers ByJG\Util\XmlUtil::innerText
-     * @todo   Implement testInnerText().
-     */
     public function testInnerText()
     {
         $dom = XmlUtil::createXmlDocumentFromStr('<root><a><item arg="1"/><item arg="2"><b1/><b2/></item><item arg="3"/></a></root>');
@@ -346,10 +282,6 @@ class XmlUtilTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals("\n<b1/><b2/>\n", $text);
     }
 
-    /**
-     * @covers ByJG\Util\XmlUtil::innerXML
-     * @todo   Implement testInnerXML().
-     */
     public function testInnerXML()
     {
         // Remove the following lines when you implement this test.
@@ -358,10 +290,6 @@ class XmlUtilTest extends \PHPUnit_Framework_TestCase
         );
     }
 
-    /**
-     * @covers ByJG\Util\XmlUtil::copyChildNodesFromNodeToString
-     * @todo   Implement testCopyChildNodesFromNodeToString().
-     */
     public function testCopyChildNodesFromNodeToString()
     {
         // Remove the following lines when you implement this test.
@@ -370,10 +298,6 @@ class XmlUtilTest extends \PHPUnit_Framework_TestCase
         );
     }
 
-    /**
-     * @covers ByJG\Util\XmlUtil::saveXmlNodeToString
-     * @todo   Implement testSaveXmlNodeToString().
-     */
     public function testSaveXmlNodeToString()
     {
         // Remove the following lines when you implement this test.
@@ -382,10 +306,6 @@ class XmlUtilTest extends \PHPUnit_Framework_TestCase
         );
     }
 
-    /**
-     * @covers ByJG\Util\XmlUtil::br2nl
-     * @todo   Implement testBr2nl().
-     */
     public function testBr2nl()
     {
         // Remove the following lines when you implement this test.
@@ -394,10 +314,6 @@ class XmlUtilTest extends \PHPUnit_Framework_TestCase
         );
     }
 
-    /**
-     * @covers ByJG\Util\XmlUtil::showXml
-     * @todo   Implement testShowXml().
-     */
     public function testShowXml()
     {
         // Remove the following lines when you implement this test.
@@ -406,9 +322,6 @@ class XmlUtilTest extends \PHPUnit_Framework_TestCase
         );
     }
 
-    /**
-     * @covers ByJG\Util\XmlUtil::removeNode
-     */
     public function testRemoveNode()
     {
         $dom = XmlUtil::createXmlDocumentFromStr('<root><subject>Text</subject><a/><b/></root>');
@@ -428,10 +341,6 @@ class XmlUtilTest extends \PHPUnit_Framework_TestCase
 
     }
 
-    /**
-     * @covers ByJG\Util\XmlUtil::removeTagName
-     * @todo   Implement testRemoveTagName().
-     */
     public function testRemoveTagName()
     {
         // Remove the following lines when you implement this test.
@@ -440,9 +349,6 @@ class XmlUtilTest extends \PHPUnit_Framework_TestCase
         );
     }
 
-    /**
-     * @covers ByJG\Util\XmlUtil::xml2Array
-     */
     public function testXml2Array1()
     {
         $xml = XmlUtil::createXmlDocumentFromFile(__DIR__ . '/buggy.xml');
@@ -451,9 +357,6 @@ class XmlUtilTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals([ "node" => [ "subnode" => "value"]], $array);
     }
 
-    /**
-     * @covers ByJG\Util\XmlUtil::xml2Array
-     */
     public function testXml2Array2()
     {
         $xml = XmlUtil::createXmlDocumentFromStr('<root><node param="pval">value</node></root>');
