+++ tests/buggy.xml
@@ -1,0 +1,5 @@
+﻿<root>
+    <node>
+        <subnode>value</subnode>
+    </node>
+</root>
