--- tests/src/buggy.xml
@@ -1,5 +1,0 @@
-﻿<root>
-	<node>
-		<subnode>value</subnode>
-	</node>
-</root>
